Break apart integration tests for "add" and "check" subcommands
