+++ Cargo.toml
@@ -18,5 +18,6 @@ tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 [dev-dependencies]
 assert_cmd = { version = "2.0.14", features = ["color"] }
 ctor = "0.2.7"
+lazy_static = "1.4.0"
 predicates = "3.1.0"
 tempfile = "3.10.1"
